#588 catching exceptions in rate limiter thread, fixed one metric, added logging

+++ hermes-consumers/src/main/java/pl/allegro/tech/hermes/consumers/consumer/rate/maxrate/MaxRateProviderFactory.java
@@ -38,7 +38,7 @@ public class MaxRateProviderFactory {
                 providerCreator = (subscription, sendCounters) -> {
                     String consumerId = configFactory.getStringProperty(CONSUMER_WORKLOAD_NODE_ID);
                     int historyLimit = configFactory.getIntProperty(CONSUMER_MAXRATE_HISTORY_SIZE);
-                    double fallbackMaxRate = configFactory.getIntProperty(CONSUMER_MAXRATE_MIN_MAX_RATE);
+                    double fallbackMaxRate = configFactory.getDoubleProperty(CONSUMER_MAXRATE_MIN_MAX_RATE);
 
                     return new NegotiatedMaxRateProvider(consumerId, maxRateRegistry, maxRateSupervisor,
                             subscription, sendCounters, metrics, fallbackMaxRate, historyLimit);
